cleaned up title with regex and added preview mode

+++ src/components/Posts.jsx
@@ -5,6 +5,7 @@ import moment from "moment";
 const Posts = () => {
   const [posts, setPosts] =  useState(null)
   const [authenticated, setAuthenticated] = useState(false)
+  const [preview, setPreview] = useState(false)
 
   useEffect(() => {
     const fetchPosts = async () => {
@@ -13,7 +14,7 @@ const Posts = () => {
           Authorization: `Bearer ${localStorage.getItem("token")}`
         }
       });
-      const { posts, current_user: currentUser } = await response.json();
+      let { posts, current_user: currentUser } = await response.json();
       setPosts(posts);
       if (currentUser) {
         setAuthenticated(true)
@@ -49,10 +50,9 @@ const Posts = () => {
     return bool ? "Public" : "Private"
   }
 
-  const logout = (e) => {
+  const togglePreview = (e) => {
     e.preventDefault()
-    localStorage.removeItem("token")
-    window.location.reload()
+    setPreview((state) => !state)
   }
 
   const onDeleteLinkClick = async (e, id) => {
@@ -73,41 +73,58 @@ const Posts = () => {
     setPosts(updatedPosts);
   }
 
+  const createTitle = (body) => {
+    let title = body.split("\n")[0]
+    const titleLen = title.split(" ").length
+    if (titleLen > 8) {
+      const regex = /^(?:\S+\s+?){1,8}/
+      title = title.match(regex)
+    }
+    return title
+  }
+
   return (
     posts && (
-      <div>
-        {authenticated && (
-            <>
-              <Link to="/"><button style={{width: "100px", height: "40px", marginRight: "10px"}}>Write</button></Link>
-              <Link to="/posts"><button style={{width: "100px", height: "40px"}} onClick={logout}>Logout</button></Link>
-            </>
-          )
-        }
-        {posts.map((post, index) => {
-          if (authenticated || post.public) {            
-            return (
-              <div key={index}>
-                <h3>
-                  <Link to={`/posts/${post.id}`}>
-                    {post.body.substring(0, 15)}
-                  </Link>
-                </h3>
-                <p>{moment(post.created_at).format("MMMM Do YYYY, h:mm:ss a")}</p>
-                {authenticated && (
-                  <>
-                    <Link to="/" className="post-links" onClick={(e) => togglePrivacy(e, post)}>{renderPrivacy(post.public)}</Link>
-                    <Link to={`/posts/${post.id}/edit`} className="post-links">Edit</Link>
-                    <Link to={`/posts`} className="post-links" onClick={(e) => onDeleteLinkClick(e, post.id)}>Delete</Link>
-                  </>
-                )}
-                <hr/>
-              </div>
-            );
-          } else {
-            return null
+      <>
+        {((!authenticated || preview) && (
+          <div style={{ border: "2px solid black", padding: "10px"}}>
+            <p>Hi, I'm Harrison. I currently work at <a href="https://coderacademy.edu.au/">CoderAcademy</a> where I've mentored around 100 students leading classes and assisting with content development. Here are my current <Link to="/projects">projects</Link>.</p>
+          </div>
+        ))}
+        <div>
+          {(authenticated && !preview) && (
+              <>
+                <Link to="/"><button style={{width: "100px", height: "40px", marginRight: "10px"}}>Write</button></Link>
+                <Link to="/posts"><button style={{width: "100px", height: "40px"}} onClick={togglePreview}>Preview</button></Link>
+              </>
+            )
           }
-        })}
-      </div>
+          {posts.map((post, index) => {
+            if ((authenticated && !preview) || post.public) {            
+              return (
+                <div key={index}>
+                  <h3>
+                    <Link to={`/posts/${post.id}`}>
+                      {createTitle(post.body)}
+                    </Link>
+                  </h3>
+                  <p>{moment(post.created_at).format("MMMM Do YYYY, h:mm:ss a")}</p>
+                  {(authenticated && !preview) && (
+                    <>
+                      <Link to="/" className="post-links" onClick={(e) => togglePrivacy(e, post)}>{renderPrivacy(post.public)}</Link>
+                      <Link to={`/posts/${post.id}/edit`} className="post-links">Edit</Link>
+                      <Link to={`/posts`} className="post-links" onClick={(e) => onDeleteLinkClick(e, post.id)}>Delete</Link>
+                    </>
+                  )}
+                  <hr/>
+                </div>
+              );
+            } else {
+              return null
+            }
+          })}
+        </div>
+      </>
     )
   );
 };
